Code maintenance: Fix some GCC warnings caused by mismatched signedness comparisons in the test assertions
Macros like HART_EQ(), HART_GT() etc have different type templates for LHS and LHS (as they should have), so plain numbers resolve to a signed int. Using a unit literal is the best thing, but explicit unsigned literals are also fine. Maybe I'll add "_pcs" or "_pieces" literal at some point.

+++ tests/envelope_contract_checker.hpp
@@ -79,7 +79,7 @@ private:
     bool m_prepareCalled = false;
 
     std::thread::id m_threadIdWhereInstanceWasConstructed {};
-    int m_numRenderNextBlockCalls = 0;
+    size_t m_numRenderNextBlockCalls = 0;
     size_t m_numFramesRendered = 0;
 
     void verify() const
